pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[tool.poetry]
name = "answer-correctly-to-go-further-telegram-bot"
version = "0.0.1"
description = ""
authors = ["Ivan Reshetnikov <resivalex@gmail.com>"]
packages = [{ include = "*", from = "src" }]

[tool.poetry.dependencies]
python = "~=3.10"
python-telegram-bot = "~=13.7"
